+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Fixed
 
+- Fix $filter for navigation elements 
+
 ## Version 1.4.25 - 2020-04-08
 
 ### Fixed